# biblioteca-digital/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Biblioteca Digital API",
    "version": "1.0.0",
    "description": "API REST da Biblioteca Digital com autenticação JWT e documentação Swagger"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Servidor local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Usuario": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "nome": { "type": "string" },
          "email": { "type": "string" },
          "senha": { "type": "string" },
          "tipo_usuario": { "type": "string" },
          "cpf": { "type": "string" },
          "telefone": { "type": "string" }
        },
        "required": ["nome", "email", "senha", "cpf"]
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/usuarios": {
      "get": {
        "summary": "Lista todos os usuários",
        "responses": {
          "200": {
            "description": "Lista de usuários",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Usuario" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Cria um novo usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Usuario" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário criado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Usuario" }
              }
            }
          }
        }
      }
    },
    "/api/usuarios/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "string" }
        }
      ],
      "get": {
        "summary": "Busca usuário por ID",
        "responses": {
          "200": {
            "description": "Usuário encontrado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Usuario" }
              }
            }
          },
          "404": { "description": "Usuário não encontrado" }
        }
      },
      "put": {
        "summary": "Atualiza um usuário existente",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Usuario" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário atualizado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Usuario" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Exclui um usuário",
        "responses": {
          "204": { "description": "Usuário excluído" }
        }
      }
    },
...
  }
}